Adding read_dir/1 and read_dir!/1 to SFTPClient
Also fixing that `ssh_sftp`'s `readdir/2` can return `:eof`,
see how `ssh_sftp.erl` uses it internally for `do_list_dir/4`:

https://github.com/erlang/otp/blob/ae8c2abc757558b9836f5373da34d5f0c245a040/lib/ssh/src/ssh_sftp.erl#L757-L767

Not sure if it's better to add `:eof` as a valid return value to `may_bang!/1`
or if it's nicer to wrap it in `{:ok, :eof}` in `ReadDir`.

The comitted version is closest to the erlang one, which I think makes the most sense.